Brave Today: use profile prefs, and remove feed from cache when user clears history
Profile prefs are now used for saving publisher / source enabled or disabled. This is more future proof as this data is not ephemeral.
Also provides a method for resetting those prefs from the NTP "customize" panel. Now that this is a regular profile pref, this could also be cleared from the regular Settings page.
Sets up an event flow from clearing history to the background process which caches the feed. It's important to clear the feed cache when the history is cleared since the feed assigns higher rank to content from hosts that the user has previously visited.
Also allows Today to be turned on and off.